% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropper.R
\name{cropper}
\alias{cropper}
\title{Cropper Widget}
\usage{
cropper(url, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{url}{path or url to an image}

\item{width}{widget width in pixels}

\item{height}{widget height in pixels}

\item{elementId}{see \link[htmlwidgets:createWidget]{htmlwidgets::createWidget}}
}
\description{
Wrapper for \href{https://github.com/fengyuanchen/cropperjs}{cropper.js}. Shows
image and lets user select a subarea within the image.
}
