% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\name{img}
\alias{img}
\title{Plain Image}
\usage{
img(url, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{url}{path or url to an image}

\item{width}{widget width in pixels}

\item{height}{widget height in pixels}

\item{elementId}{see \link[htmlwidgets:createWidget]{htmlwidgets::createWidget}}
}
\description{
HTML widgets that creates simple \code{&lt;img&gt;} html tag.
}
